Downgrade Apache Livy to 0.5.0
Apache Livy > 0.5.0 requires that the name of the session be unique, and
will throw a 400 if that is not the case. EMR does not have a
requirement of unique step names, and so its behaviour is more closely
mimicked by earlier Livy versions. The plan is to ideally move away from
using Livy as a backend, so this is considered temporary anyway.