Add preliminary S3 integration commit (#6)
* Add preliminary S3 integration commit

Under the current design the spark-submit call is altered and the
S3 paths changed to the local path. Consider leveraging Moto to have the
job run in a simulated S3 environment, where all the necessary data has
been already been 'uploaded' to the in memory S3. The question of how to
control the two sets of S3 credentials is a bit tricky though, but the
general idea seems less hacky.